Add support for configuration file (#213)
This PR adds support for the configuration file as proposed https://github.com/sbdchd/squawk/issues/205

### Feature
This change prioritizes the CLI flag before it starts using the config file.

The file name is `.squawk.toml` and if a custom path is not specified, then Squawk tries to look up the directory tree to find a configuration file. If not found, the program continues running without a config file.

### Testing
These changes have been tested only on macOS, it would be great if someone wants to test these on Linux, but it should work there as well.