Fix flake version (#256)
Hello! I noticed when using the flake that the Nix default package version didn't match what `squawk --version` showed. I found that the version was out of date in `flake.nix` so I updated that and the `s/update-version` script so that it should be kept in sync for the future.

Awesome project, by the way!