Improve constraint_missing_not_valid attribution
Prior to this, there would be two errors reported if there was an
explicit COMMIT and assume_in_transaction was set to true. I failed to
exclude the second error when there was an implicit transaction and an
explicit COMMIT. Probably not a common case, but wanted to address it
all the same.