Update Cargo.lock to relect version update in 9afce6a (#266)
I was noticing that Cargo.lock had unstaged changes in it when I built against a recent version of master. It looks like the version was bumped in 9afce6a for cli/Cargo.toml, but the lock file didn't get updated. I'm assuming this should be changed as well, but feel free to close if that's not the case.