Add flag to inform Squawk that .sql file(s) will be wrapped in a transaction (#264)
This addresses #260 by adding a new CLI option called `--assume-in-transaction` that can be supplied to Squawk to indicate that an external SQL migration tool will wrap any SQL files in a transaction when run. Squawk can then use this information to accurately apply its linting rules. By default, assume_in_transaction is false, continuing the current behavior for Squawk.

This can also be set in the .squawk.toml config file as well with `assume_in_transaction = true`. The CLI flag overrides this, as is the case with other options.

`assume_in_transaction` is now piped through to each lint rule, in a similar manner to `pg_version`. This required updating all rule functions to accept this new parameter, even if it was unused. I think this could be improved by passing a (`RuleOptions` or similarly named) struct to each rule as opposed to individual parameters. I'm happy to make that change in a subsequent PR, but didn't want to intermix that with this change. 

There were some cases where rules should make use of `assume_in_transaction`. I also updated the linting code and wrote tests for each rule to ensure that the behavior is correct.

I'm happy to rework almost any of this, so please give feedback.