Recognize most DROP statements within a transaction as robust (#276)
PostgreSQL only has a handful of DROP statements that may not be executed in a transaction. These include DROP DATABASE, DROP TABLESPACE, and sometimes DROP SUBSCRIPTION, according to my reading of the manual. These variants are parsed to different enum values in the AST and won't fall under `DropStmt` here.

Otherwise, DROP statements are allowed in transactions and can be rolled back with the transaction, leaving the database unchanged. This robustness-checking rule should not flag DROP statements in a transaction. Add new test cases to ensure this behavior as well.