Use libpg_query version 13-2.1.0 instead of the latest in nixpkgs
Squawk isn't able to handle the newest 15-* version of libpg_query that
is present in nixpkgs now. Because it is extracted from a newer version
of PostgreSQL, I believe the abstract syntax tree that it generates is
slightly different and Squawk isn't expecting those differences. For
that reason, pin libpg_query to 13-* until that is addressed in Squawk.
13-2.1.0 comes from what libpg_query-sys, the Rust bindings to
libpg_query is using.

The issue requiring the `allowUnsupportedSystem` directive was fixed as
part of https://github.com/NixOS/nixpkgs/pull/179749 and is now no
longer necessary.