Warn when adding a required field to an existing table
This was extracted from adding-not-null-field into a new rule. When
support for altering rules based on the PostgreSQL version was
introducted in 4624df3dafb184b295216b8e47db3b22d98c8769,
adding-not-null-field was changed in a way that caused this check to be
skipped for newer versions of PostgreSQL. I think that change was valid
for the documented intent of adding-not-null-field, but generally, it is
still problematic to add a required field to an existing table. This new
rule reinstates that check for all PostgreSQL versions and clarifies
what the problem is.