Change behavior of adding-not-nullable-field w.r.t. added columns
Previously, this rule was looking for NOT NULL columns without a
DEFAULT. I can't figure out why that is and think it might have been a
mistake. When I read the docs written for this rule, the examples show
this check flagging

```
ALTER TABLE "recipe" ADD COLUMN "view_count" integer DEFAULT 10 NOT NULL;
```

which has both NOT NULL and DEFAULT constraints. Arguably, this could be
relaxed to only consider a NOT NULL constraint when applying the
violation. However, the new adding-required-field rule covers the
situation when adding a NOT NULL column without a DEFAULT. It seems most
appropriate then that this rule is applied when there is a DEFAULT, but
only for PostgreSQL < 11 (which is already a part of this rule).

Tests are updated accordingly.