Always comment when Squawk checks a file
Previously we would only add or update a comment if violations found. In cases where violations were resolved, this would leave stale warnings in comments.

Now we will always comment as long as files are found and checked.

There is a slight edge case where if a file is no longer in the pull request, the comment won't be updated.