Issue-338: Adjust handling of ban-concurrent-index-creation-in-transaction (#339)
This changes the logic of `ban-concurrent-index-creation-in-transaction` when `assume_in_transaction` is set. Tools like `golang-migrate` always wrap migrations in transactions but support individual statements like `CREATE INDEX CONCURRENTLY` by not wrapping them in transactions.

To support this, this commit alters the logic of `ban-concurrent-index-creation-in-transaction` to allow a standalone index create command even when `assume_in_transaction` is `true`.