Ignore postgres version in adding-not-nullable-field (#412)
### Summary
Hey there, I noticed in https://github.com/sbdchd/squawk/pull/301 that the logic for handling adding new fields as null was split out from [adding-not-nullable-field](https://squawkhq.com/docs/adding-not-nullable-field) into [adding-required-field](https://squawkhq.com/docs/adding-required-field).

The logic for performing the pg version check wasn't removed however, so this causes the rule to fail to detect changing existing columns to be not null. All the logic for checking not null for adding cols was extracted to `adding-required-field` or covered by `adding-required-field`, so I think we can remove the version check altogether.

### Examples
https://github.com/sbdchd/squawk/pull/301 has nice examples showing how the other rules cover the adding column with not null. I've included the part below to show what changes with this PR.

Previously with --pg-version >=12:
```SQL
ALTER TABLE foo ALTER COLUMN d SET NOT NULL; -- no warning,
```

This PR with --pg-version >=12:
```SQL
ALTER TABLE foo ALTER COLUMN d SET NOT NULL; -- triggers adding-not-nullable-field
```

### Documentation
I updated `adding-not-nullable-field` to explicitly mention that it no longer covers adding cols, and added to the problem+solution sections as well.

I think ideally it'd be best to maybe re-name this rule to clarify the fact that it only covers modifying existing cols now, but that may be hard to do

---
Please let me know if this change may not be the best approach, or if there are any concerns!