internal: refactor config (#681)
Right how we handle the various configuration knobs is a mess.

We basically need to resolve the cli options and the file options into structs for each of the possible commands to execute.

The enum for the github actions is nice, but gets a little messy here.

We're also missing parity between the file config and the cli options. Some options are in the cli but not in the file.

There's a lot of room for improvement.

I want to get this sorted to make https://github.com/sbdchd/squawk/issues/679 easier to fix.